+++ src/main/java/com/intellij/plugins/haxe/ide/refactoring/extractMethod/UnableToExtractMethodException.java
@@ -1,0 +1,4 @@
+package com.intellij.plugins.haxe.ide.refactoring.extractMethod;
+
+public class UnableToExtractMethodException  extends  RuntimeException{
+}